requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cryptography==41.0.7
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
anthropic==0.39.0
